Code that should have been on the develop branch ages ago, but I was too lazy.
Commits over this period:
- Made a basic implementation of pybind and displaying the build date
- Updated build files
- not working due to compile error- trying to pass singleton
- desperation to get singletons to work
- dir(inst)
- Reverted code
- Successfully exposed Vector2D to python as well as fixing executing Python from files (added line endings.)
- Cleaned up code but nothing is rendering
- Fixed drawing by reverting to older code - also added script loading support to states
- Added more binds to WarspiteObject as well as fixing build problems
- Added more methods exposed to Python as well as updated the test script
- Added some additional string utilities for file paths - level objects now carry their path and report their name
- Made some data structures to expose more functionality to Python - testing by assigning level pointer at runtime.
- Made progress - access violation when assigning level pointer
- Added code to expose more functionality to Python as well as refining it
- Added Game object bindings (still borked)
- Added script bindings for CWarspiteUtil as well as modifying the naming to match the naming scheme
- Removed boost include, SoundManager init + included old test script
- Errors during Python execution are now reported + small changes
- Made the ScriptManager load before states are loaded and objects do OnPlay on the mainmenu
- Updated some bindings
- Created GameStateDictionary class
- Made resolving filepaths centralised
- Centralised how JSON is loaded from disk
- Finished script bindings for the Game Object
- Some refactoring of names and case sens parts
- Added missing headers
- Using ctime from the C instead of MS non portable version
- Objects can now be created in Python
- Access violation - prolly due to the original pointer being destroyed
- THIS!!! DOES!!! NOT!!! WORK!!!
- Made a wrapper for Python to interact with objects created in C++
- Major overhall of where code is located (mostly script)
- Fixed access violation caused by the interpreter not initialising correctly
- New expriemental buildscript
- Updated cache path
- Updated CMake and build files
- Updated script
- Updated script
- Updated Buildscript
- Removed dogey script
- Made a new variable for whether objects should overlap
- Added logging to the Engine
- Python now outputs directly to the log file
- Made script output more readable in logs
- Beginning of implmentation of the FontManager
- Borked remove file name code
- Basic Implementation of FontManager and bugfixes

+++ Engine/GameStateBase.cpp
@@ -17,12 +17,12 @@ bool CGameStateBase::OnEnd()
 	m_shouldDraw = false;
 
 	// Destroying everything!
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{
 		m_GameObjects[i]->Destroy();
 	}
 
-	for (int i = 0; i < m_TextureIDList.size(); i++)
+	for (size_t i = 0; i < m_TextureIDList.size(); i++)
 	{
 		CTextureManager::Instance()->Remove(m_TextureIDList[i]);
 	}
@@ -36,7 +36,7 @@ bool CGameStateBase::OnEnd()
 void CGameStateBase::OnThink()
 {
 	if (!m_shouldTick) return;
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{	
 		if (!m_GameObjects[i]->OnThink()) return;
 	}
@@ -46,7 +46,7 @@ void CGameStateBase::Draw()
 {
 	if (!m_shouldDraw) return;
 
-	for (int i = 0; i < m_GameObjects.size(); i++)
+	for (size_t i = 0; i < m_GameObjects.size(); i++)
 	{
 		m_GameObjects[i]->Draw();
 	}
